% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdbootstraps.R
\name{rdboot}
\alias{rdboot}
\title{The main function to perform bootstraps in sharp regression discontinuity
design.}
\usage{
rdboot(y, x, c = 0, p = 1, q = 2, h = NULL, b = NULL, Nbc = 500,
  Nci = 999, kernel = "uni", level = c(90, 95, 99))
}
\arguments{
\item{y}{is the dependent variable.}

\item{x}{is the running variable (score or forcing variable).}

\item{c}{specifies the RD cutoff in x; default is c = 0.}

\item{p}{specifies the order of the local-polynomial used to construct the}

\item{q}{specifies the order of the local-polynomial used to construct the
bias-correction; default is q = 2 (local quadratic regression).}

\item{h}{directly sets the main bandwidth. If not specified, it is computed
by the command rdrobust::rdbwselect.}

\item{b}{directly sets the pilot bandwidth. If not specified, it is computed
by the command rdrobust::rdbwselect.}

\item{Nbc}{is the number of bootstraps for bias correction.}

\item{Nci}{is the number of bootstraps to construct percentile confidence
interval.}

\item{kernel}{is the kernel function used to construct the local-polynomial
estimator(s). Options are triangular (default option), epanechnikov and
uniform.}

\item{level}{sets the confidence levels, a vector of length 3.}

\item{p}{specifies the order of the local-polynomial used to construct the
point-estimator; default is p = 1 (local linear regression).}
}
\value{
A matrix containing point and interval estimates.
}
\description{
This package was created to speed up the simulations using
multiple cores. Functions compilded from C++ code cannot be loaded to each
worker unless they are bundled into a package. This package is primarily only
for sharp RD simulations in this paper and is not designed for general use.
}

